chore: update run local server script (#1180)

+++ script/run_local_server.sh
@@ -9,6 +9,11 @@ DB_PASSWORD="${POSTGRES_PASSWORD:=password}"
 DB_PORT="${POSTGRES_PORT:=5432}"
 DB_HOST="${POSTGRES_HOST:=localhost}"
 
+if [ ! -f .env ]; then
+    echo ".env file not found in the current directory. Try: cp dev.env .env"
+    exit 1
+fi
+
 # Stop and remove any existing containers to avoid conflicts
 docker compose --file ./docker-compose-dev.yml down
 
@@ -47,13 +52,20 @@ done
 # To generate the .sqlx files, we need to run the following command
 # After the .sqlx files are generated, we build in SQLX_OFFLINE=true
 # where we don't need to connect to the database
+# for example: SQLX_OFFLINE=true ./script/run_local_server.sh
 cargo sqlx database create && cargo sqlx migrate run
 if [[ -z "${SKIP_SQLX_PREPARE+x}" ]]
 then
   cargo sqlx prepare --workspace
 fi
 
+
+# To skip this step, set SKIP_APPFLOWY_CLOUD=true before running the script, like this:”
+# SKIP_APPFLOWY_CLOUD=true ./script/run_local_server.sh. By default, this step is not skipped.
+if [[ -z "${SKIP_APPFLOWY_CLOUD+x}" ]]
+then
 cargo run --package xtask
+fi
 
 # revert to require signup email verification
 export GOTRUE_MAILER_AUTOCONFIRM=false
